[issue #156] 대댓글 있는 댓글 삭제 로직 변경

See merge request s06-webmobile2-sub2/S06P12D106!160

+++ backend/src/main/java/com/sweet/home/comment/controller/dto/response/CommentReplyResponse.java
@@ -30,13 +30,13 @@ public class CommentReplyResponse {
     private LocalDateTime updatedAt;
 
     @JsonProperty("total_likes")
-    private long totalLikes;
+    private int totalLikes;
 
     protected CommentReplyResponse() {
     }
 
     public CommentReplyResponse(Long id, Long parentId, String username, String content, LocalDateTime createdAt,
-        LocalDateTime updatedAt, long totalLikes) {
+        LocalDateTime updatedAt, int totalLikes) {
         this.id = id;
         this.parentId = parentId;
         this.username = username;
